+++ udata/tests/api/test_base_api.py
@@ -8,7 +8,7 @@ from udata.forms import Form, fields
 
 from . import APITestCase
 
-ns = api.namespace('fake', 'A Fake namespace')
+ns = api.namespace('fake-ns', 'A Fake namespace')
 
 
 @ns.route('/options', endpoint='fake-options')